% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bills_data.R
\name{get_bill_data}
\alias{get_bill_data}
\title{Obtener datos crudos de La Década Votata
(\emph{Download bill raw data})}
\usage{
get_bill_data(chamber = NULL)
}
\arguments{
\item{chamber}{parámetro para elegir entre las cámaras de 'diputados' o 'senadores'}
}
\description{
Función que devuelve una \emph{list} con las bases de datos crudas disponible
}
\seealso{
\link{show_available_bills}
}
