% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bill_result.R
\name{get_bill_result}
\alias{get_bill_result}
\title{Obtiene el resultado de la votación de un proyecto de leu}
\usage{
get_bill_result(bill = NULL)
}
\arguments{
\item{bill}{identificador del proyecto de ley que se obtiene con \link{show_available_bills}}
}
\description{
Obtiene el resultado de la votación de un proyecto de leu
}
