% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codebook.R
\name{get_codebook}
\alias{get_codebook}
\title{Obtener libro de códigos de bases crudas del proyecto la 'Década Votada'}
\usage{
get_codebook(data = NULL)
}
\arguments{
\item{data}{define el libro de códigos a desargar del conjunto de conjunto de datos disponible}
}
\value{
a "tbl_df" "tbl" "data.frame" object
}
\description{
Obtener libro de códigos de bases crudas del proyecto la 'Década Votada'
}
