% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Obtener datos crudos de La Década Votata
(\emph{Download bill raw data})}
\usage{
get_raw_data(chamber = NULL, base = NULL)
}
\arguments{
\item{chamber}{parámetro para elegir entre las cámaras de 'diputados' o 'senadores'}

\item{base}{parámetro para definir que base de datos descargar. Las opciones son  'bloques', 'votaciones', 'asuntos' y  'diputados'. El valor por defecto NULL devuelve las tres bases disponibles.}
}
\description{
Funcion que devuelve una \emph{list} con las bases de datos crudas disponible
}
\seealso{
\link{show_available_bills}
}
