% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislAr.R
\docType{package}
\name{legislAr}
\alias{legislAr}
\title{\code{legislAr} package}
\description{
Caja de Herramientas para el trabajo con datos legislativos de Argentina
See the README on
\href{https://github.com/PoliticaArgentina/legislAr/README.md}{Github}
}
