% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_available_bills.R
\name{show_available_bills}
\alias{show_available_bills}
\title{Diccionario de proyectos de ley sometidos a votación
(\emph{Bills collection})}
\usage{
show_available_bills(chamber = NULL, viewer = FALSE)
}
\arguments{
\item{chamber}{parámetro para elegir entre las cámaras de 'diputados' o 'senadores'}

\item{viewer}{Por default es \code{TRUE} y  muestra una tabla formateada en el \emph{Viewer} de \emph{RStudio}. Cuando \code{FALSE} imprime en consola.}
}
\description{
Función que devuelve un \emph{data.frame} listado de proyectos de ley con un id por proyecto y cámara legislativa
}
\examples{

show_available_bills(chamber = 'senadores' , viewer = FALSE)


}
\seealso{
\link{get_bill_votes}
}
